Update mods for GlassSoldier
She requested some and I'm happy to oblige:
- fossilsarcheology

These I added because I like 'em:
- atum2
- arcane archives

Had to disable Bamboozled because of a roots conflict; disabled
Modular Machinery because it was causing problems and we never did get
around to acutally using it.  Ditched Jack O' Launcher because I never
used it.  Lost wards because we ran out of potion IDs.